add convenience scripts to dotfiles
The main purpose is to setup a local worktree of a repo given in
scr/bare_git/setupWorktree.sh.
Git hooks are not included, as they are not customizable per repository.

Other scripts are nix helpers for experimental purpose.
I am still dissatisfied with the tutorials and performance of how to
handle nix flakes out-of-tree, so I do not commit yet to a nix
configuration.